+++ riverpie/lib/src/util/stacktrace.dart
@@ -1,0 +1,8 @@
+Never rethrowWithNewStackTrace(Object error, StackTrace stackTrace) {
+  Error.throwWithStackTrace(
+    error,
+    StackTrace.fromString(
+      '$stackTrace===== asynchronous gap ===========================\n${StackTrace.current}',
+    ),
+  );
+}