+++ riverpie/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.22.0
 
+- fix: should emit events within `init` of a `ReduxNotifier`
 - feat: make add-on actions customizable
 - **BREAKING**: events now have `millisSinceEpoch` instead of `microsSinceEpoch`